refactor: extract MessageListView into its own file
